Revert "Move big parts of cmd_log_send into libkb"
This reverts commit 14caa41c11440bb77fd1802f8847931c086e7896.
